+++ src/shaders/sparkle/fs_stretched_sprite.glsl
@@ -6,19 +6,18 @@
 
 // ----------------------------------------------------------------------------
 
-layout(location = 0) out vec4 fragColor;
-
 in GDataBlock {
   vec3 color;
   vec2 texcoord;
   float decay;
 } IN;
 
+layout(location = 0) out vec4 fragColor;
+
 // ----------------------------------------------------------------------------
 
 void main() {
-  fragColor = compute_color(IN.color, IN.texcoord);
-  fragColor *= IN.decay * 0.25;
+  fragColor = compute_color(IN.color, IN.decay, IN.texcoord);
 }
 
 // ----------------------------------------------------------------------------