+++ src/ui/views/Rendering.cc
@@ -20,17 +20,34 @@ void Rendering::render() {
   
   ImGui::Combo("Mode", reinterpret_cast<int*>(&params_.rendermode),
     kRenderModeDescriptions, IM_ARRAYSIZE(kRenderModeDescriptions));
-  if (GPUParticle::STRETCHED == params_.rendermode) {
-    ImGui::DragFloat("Stretch factor", &params_.stretched_factor,
-      kStretchedFactorStep, kStretchedFactorMin, kStretchedFactorMax);
+  switch (params_.rendermode) {
+    case GPUParticle::STRETCHED:
+      ImGui::DragFloat("Stretch factor", &params_.stretched_factor,
+        kStretchedFactorStep, kStretchedFactorMin, kStretchedFactorMax);
+      ImGui::Separator();
+    break;
+
+    case GPUParticle::POINTSPRITE:
+    default:
+      ImGui::DragFloatRange2("Particle size", &params_.min_size, &params_.max_size,
+        kParticleSizeStep, kParticleSizeMin, kParticleSizeMax, "Min: %.2f", "Max: %.2f");
+      ImGui::Separator();
+    break;
   }
   
   ImGui::Combo("Color", reinterpret_cast<int*>(&params_.colormode),
     kColorModeDescriptions, IM_ARRAYSIZE(kColorModeDescriptions));
   if (GPUParticle::GRADIENT == params_.colormode) {
+
+    ImGui::Spacing();
     ImGui::ColorPicker4("Birth gradient", params_.birth_gradient);
+    ImGui::Spacing();
     ImGui::ColorPicker4("Death gradient", params_.death_gradient);
   }
+  ImGui::Separator();
+
+  ImGui::DragFloat("Fade coefficient", &params_.fading_factor,
+    kFadingFactorStep, kFadingFactorMin, kFadingFactorMax);
 }
 
 }  // namespace views