+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## [Unreleased]
 ### Added
+- User Interface with ImgGUI with Simulation, Rendering, and Debug parameters.
 - Submodules dependencies : GLFW, GLM and imgui.
 - This changelog.
 
 ### Changed
 - Improve CMake build overall. Switch to C++14.
+- Some internal simulations parameters have slightly changed, a lot have been set to be controlled through the user interface.
 - Fixes C-style cast and type conversions.
 
 ### Removed